Moved test submodule to webdriver.support subpackage
Updated references to it and example project
