+++ webdriver_test_tools/__init__.py
@@ -1,7 +1,6 @@
 from . import config
 from . import data
 from . import project
-from . import test
 from . import webdriver
 from . import common
 from . import cmd
