+++ webdriver_test_tools/webdriver/support/__init__.py
@@ -1,1 +1,2 @@
 from . import expected_conditions
+from . import test
