+++ webdriver_test_tools/webdriver/support/test.py
@@ -6,7 +6,6 @@ from selenium.common.exceptions import TimeoutException, NoSuchElementException
 from selenium.webdriver.support import expected_conditions as EC
 
 # TODO: consistent naming conventions
-# TODO: move to webdriver.support subpackage
 # TODO: ensure WebDriverTestCase has assertion methods for all of these
 
 
@@ -64,6 +63,7 @@ def is_scrolled_into_view(driver, element, fully_in_view=True):
 
 
 # Expected Condition Tests
+# TODO: extract these to another module?
 
 def expected_condition_test(driver, ec_object, wait_timeout=10):
     """Test for an expected condition until wait timeout is reached
